+++ src/main/java/ch/wisv/events/controller/rest/EventRESTController.java
@@ -3,8 +3,8 @@ package ch.wisv.events.controller.rest;
 import ch.wisv.events.data.model.event.Event;
 import ch.wisv.events.data.model.event.EventStatus;
 import ch.wisv.events.data.model.product.Product;
-import ch.wisv.events.response.EventsDefaultResponse;
-import ch.wisv.events.service.EventService;
+import ch.wisv.events.response.event.EventsDefaultResponse;
+import ch.wisv.events.service.event.EventService;
 import ch.wisv.events.utils.ResponseEntityBuilder;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
